// src/app/shared/models/country.model.ts
export interface Country {
    name: string;
    topLevelDomain: string[];
    alpha2Code: string;
    alpha3Code: string;
    callingCodes: number[];
    capital: string;
    altSpellings: string[];
    region: string;
    subregion: string;
    population: number;
    latlng: number[];
    demonym: string;
    area: number;
    gini: string;
    timezones: string[];
    borders: string[];
    nativeName: string;
    numericCode: number;
    currencies: [{
        code: string;
        name: string;
        symbol: string;
    }];
    languages: [{
        iso639_1: string;
        iso639_2: string;
        name: string;
        nativeName: string;
    }];
    translations: {
        de: string;
        es: string;
        fr: string;
        ja: string;
        it: string;
        br: string;
        pt: string;
        nl: string;
        hr: string;
        fa: string;
    };
    flag: string;
    regionalBlocs: [{
        acronym: string;
        name: string;
        otherAcronyms: string[]
        otherNames: string[]
    }];
    cioc: string;
}